scripts/requirements.txt
pyelftools
tabulate
